refactor: move components into appropriate folders
