+++ src/components/layout/Footer.tsx
@@ -11,7 +11,7 @@ import {
   SiYoutube,
 } from 'react-icons/si';
 
-import UnstyledLink from '@/components/links/UnstyledLink';
+import UnstyledLink from '@/components/atoms/links/UnstyledLink';
 
 export default function Footer() {
   return (
