+++ src/components/molecules/ContactForm/ContactForm.tsx
@@ -8,10 +8,10 @@ import * as Yup from 'yup';
 
 import { verifyCaptcha } from '@/lib/verifyCaptcha';
 
+import Button from '@/components/atoms/buttons/Button';
 import { Input } from '@/components/atoms/Input/Input';
 import { Select } from '@/components/atoms/select/Select';
 import { TextArea } from '@/components/atoms/TextArea/TextArea';
-import Button from '@/components/buttons/Button';
 
 export interface ContactFormProps {
   subjects: Subject[];
